Computations involving regular matrix pencils

MatrixPencils.isregular — Function

isregular(A, E, γ; atol::Real = 0,  rtol::Real = atol1 > 0 ? 0 : n*ϵ) -> Bool

Test whether the linear pencil A-λE is regular at λ = γ(i.e., A-λE is square and det(A-γE) !== 0). The underlying computational procedure checks the maximal rank of A-γE if γ is finite and of E if γ is infinite .

The keyword arguements atol and rtol specify the absolute and relative tolerances for the nonzero elements of A-γE, respectively. The default relative tolerance is n*ϵ, where n is the size of A, and ϵ is the machine epsilon of the element type of A.

isregular(A, E; atol1::Real = 0, atol2::Real = 0, rtol::Real=min(atol1,atol2)>0 ? 0 : n*ϵ) -> Bool

Test whether the linear pencil A-λE is regular (i.e., A-λE is square and det(A-λE) !== 0). The underlying computational procedure reduces the pencil A-λE to an appropriate Kronecker-like form (KLF), which provides information on the rank of A-λE.

The keyword arguements atol1, atol2 and rtol specify the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, and the relative tolerance for the nonzero elements of A and E, respectively. The default relative tolerance is n*ϵ, where n is the size of A, and ϵ is the machine epsilon of the element type of A.

MatrixPencils.isunimodular — Function

isunimodular(A, E; atol1::Real = 0, atol2::Real = 0, rtol::Real=min(atol1,atol2)>0 ? 0 : n*ϵ) -> Bool

Test whether the linear pencil A-λE is unimodular (i.e., A-λE is square, regular and det(A-λE) == constant). The underlying computational procedure reduces the pencil A-λE to an appropriate Kronecker-like form (KLF), which provides information to check the full rank of A-λE and the lack of finite eigenvalues.

The keyword arguements atol1, atol2 and rtol specify the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, and the relative tolerance for the nonzero elements of A and E, respectively. The default relative tolerance is n*ϵ, where n is the size of A, and ϵ is the machine epsilon of the element type of A.

MatrixPencils.fisplit — Function

fisplit(A, E, B, C; fast = true, finite_infinite = false, atol1 = 0, atol2 = 0, rtol, withQ = true, withZ = true) -> (At, Et, Bt, Ct, Q, Z, ν, nf, ni)

Reduce the linear regular pencil A - λE to an equivalent form At - λEt = Q'*(A - λE)*Z using orthogonal or unitary transformation matrices Q and Z such that the transformed matrices At and Et are in one of the following block upper-triangular forms:

(1) if finite_infinite = false, then

               | Ai-λEi |   *     |
    At - λEt = |--------|---------|, 
               |    O   | Af-λEf  |

where the ni x ni subpencil Ai-λEi contains the infinite elementary divisors and the nf x nf subpencil Af-λEf, with Ef nonsingular and upper triangular, contains the finite eigenvalues of the pencil A-λE.

The subpencil Ai-λEi is in a staircase form, with Ai nonsingular and Ei nilpotent. The nb-dimensional vector ν contains the dimensions of the diagonal blocks of the staircase form Ai-λEi such that i-th block has dimensions ν[i] x ν[i]. The difference ν[i]-ν[i+1] for i = 1, 2, ..., nb is the number of infinite elementary divisors of degree i (with ν[nb+1] = 0).

(2) if finite_infinite = true, then

               | Af-λEf |   *    |
    At - λEt = |--------|--------|, 
               |   O    | Ai-λEi |

where the nf x nf subpencil Af-λEf, with Ef nonsingular and upper triangular, contains the finite eigenvalues of the pencil A-λE and the ni x ni subpencil Ai-λEi contains the infinite elementary divisors.

The subpencil Ai-λEi is in a staircase form, with Ai nonsingular and Ei nilpotent. The nb-dimensional vectors ν contains the dimensions of the diagonal blocks of the staircase form Ai-λEi such that i-th block has dimensions ν[i] x ν[i]. The difference ν[nb-j+1]-ν[nb-j] for j = 1, 2, ..., nb is the number of infinite elementary divisors of degree j (with ν[0] = 0).

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, and the relative tolerance for the nonzero elements of A and E.

The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed left orthogonal or unitary transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed right orthogonal or unitary transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

Bt = Q'*B, unless B = missing, in which case Bt = missing is returned, and Ct = C*Z, unless C = missing, in which case Ct = missing is returned .

MatrixPencils.fisplit! — Function

fisplit!(A, E, B, C, Q, Z; fast = true, finite_infinite = false, atol1 = 0, atol2 = 0, rtol, withQ = true, withZ = true) -> (ν, nf, ni)

Reduce the linear regular pencil A - λE to an equivalent form At - λEt = Q1'*(A - λE)*Z1 using orthogonal or unitary transformation matrices Q1 and Z1 such that the transformed matrices At and Et are in one of the following block upper-triangular forms:

(1) if finite_infinite = false, then

               | Ai-λEi |   *     |
    At - λEt = |--------|---------|, 
               |    O   | Af-λEf  |

where the ni x ni subpencil Ai-λEi contains the infinite elementary divisors and the nf x nf subpencil Af-λEf, with Ef nonsingular and upper triangular, contains the finite eigenvalues of the pencil A-λE.

The subpencil Ai-λEi is in a staircase form, with Ai nonsingular and Ei nilpotent. The nb-dimensional vector ν contains the dimensions of the diagonal blocks of the staircase form Ai-λEi such that i-th block has dimensions ν[i] x ν[i]. The difference ν[i]-ν[i+1] for i = 1, 2, ..., nb is the number of infinite elementary divisors of degree i (with ν[nb+1] = 0).

(2) if finite_infinite = true, then

               | Af-λEf |   *    |
    At - λEt = |--------|--------|, 
               |   O    | Ai-λEi |

where the nf x nf subpencil Af-λEf, with Ef nonsingular and upper triangular, contains the finite eigenvalues of the pencil A-λE and the ni x ni subpencil Ai-λEi contains the infinite elementary divisors.

The subpencil Ai-λEi is in a staircase form, with Ai nonsingular and Ei nilpotent. The nb-dimensional vectors ν contains the dimensions of the diagonal blocks of the staircase form Ai-λEi such that i-th block has dimensions ν[i] x ν[i]. The difference ν[nb-j+1]-ν[nb-j] for j = 1, 2, ..., nb is the number of infinite elementary divisors of degree j (with ν[0] = 0).

The reduced matrices At and Et are returned in A and E, respectively, while Q1'*B is returned in B, unless B = missing, and C*Z1, is returned in C, unless C = missing.

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, and the relative tolerance for the nonzero elements of A and E.

The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed left orthogonal or unitary transformations Q1 are accumulated in the matrix Q <- Q*Q1 if withQ = true. Otherwise, Q is unchanged. The performed right orthogonal or unitary transformations Z1 are accumulated in the matrix Z <- Z*Z1 if withZ = true. Otherwise, Z is unchanged.

MatrixPencils._qrE! — Function

_qrE!(A, E, Q, B; withQ = true)

Reduce the linear regular pencil A - λE to an equivalent form A1 - λE1 = Q1'*(A - λE) using an orthogonal or unitary transformation matrix Q1 such that the transformed matrix E1 is upper triangular. The reduction is performed using the QR-decomposition of E.

The performed left orthogonal or unitary transformations Q1 are accumulated in the matrix Q <- Q*Q1 if withQ = true. Otherwise, Q is unchanged.

Q1'*B is returned in B unless B = missing.

MatrixPencils._svdlikeAE! — Function

_svdlikeAE!(A, E, Q, Z, B, C; fast = true, atol1 = 0, atol2 = 0, rtol, withQ = true, withZ = true) -> (rE, rA22)

Reduce the linear regular pencil A - λE to an equivalent form A1 - λE1 = Q1'*(A - λE)*Z1 using orthogonal or unitary transformation matrices Q1 and Z1 such that the transformed matrices A1 and E1 are in the following SVD-like coordinate form

               | A11-λE11 |  A12  |  A13  |
               |----------|-------|-------|, 
    A1 - λE1 = |    A21   |  A22  |   0   |
               |----------|-------|-------|
               |    A31   |   0   |   0   |

where the rE x rE matrix E11 and rA22 x rA22 matrix A22 are nosingular, and E11 and A22 are upper triangular, if fast = true, and diagonal, if fast = false. The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, and the relative tolerance for the nonzero elements of A and E.

The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed left orthogonal or unitary transformations Q1 are accumulated in the matrix Q <- Q*Q1 if withQ = true. Otherwise, Q is unchanged. The performed right orthogonal or unitary transformations Z1 are accumulated in the matrix Z <- Z*Z1 if withZ = true. Otherwise, Z is unchanged.

Q1'*B is returned in B unless B = missing and C*Z1 is returned in C unless C = missing .