Manipulation of structured linear matrix pencils

MatrixPencils.sklf — Function

sklf(A, E, B, C, D; fast = true, finite_infinite = false, atol1 = 0, atol2 = 0, rtol, withQ = true, withZ = true) -> (F, G, Q, Z, νr, μr, νi, nf, νl, μl)

Reduce the structured linear pencil M - λN

           | A-λE | B | 
  M - λN = |------|---|
           |  C   | D |

to an equivalent form F - λG = Q'*(M - λN)*Z using orthogonal or unitary transformation matrices Q and Z such that the transformed matrices F and G are in the following Kronecker-like form exhibiting the complete Kronecker structure:

            |  Mr-λNr  |     *      |    *    |
            |----------|------------|---------|
  F - λG =  |    O     | Mreg-λNreg |    *    |
            |----------|------------|---------|
            |    O     |     0      |  Ml-λNl |

The full row rank pencil Mr-λNr is in a staircase form, contains the right Kronecker indices of the pencil M-λNand has the form

     Mr-λNr  = | Br | Ar-λEr |,

where Er is upper triangular and nonsingular. The nr-dimensional vectors νr and μr contain the row and, respectively, column dimensions of the blocks of the staircase form Mr-λNr such that the i-th block has dimensions νr[i] x μr[i] and has full row rank. The difference μr[i]-νr[i] for i = 1, 2, ..., nr is the number of elementary Kronecker blocks of size (i-1) x i.

The pencil Mreg-λNreg is regular, contains the infinite and finite elementary divisors of M-λN and has the form

                   | Mi-λNi |    *    |
     Mreg-λNreg  = |--------|---------|, if `finite_infinite = false`, or the form
                   |   0    |  Mf-λNf |

                   | Mf-λNf |    *    |
     Mreg-λNreg  = |--------|---------|, if `finite_infinite = true`, 
                   |   0    |  Mi-λNi |

where: (1) Mi-λNi, in staircase form, contains the infinite elementary divisors of M-λN and Ni is upper triangular and nilpotent; (2) Mf-λNf contains the infinite elementary divisors of M-λN and Nf is upper triangular and nonsingular. The ni-dimensional vector νi contains the dimensions of the square blocks of the staircase form Mi-λNi such that the i-th block has dimensions νi[i] x νi[i]. If finite_infinite = true, the difference νi[i]-νi[i+1] for i = 1, 2, ..., ni is the number of infinite elementary divisors of degree i (with νi[ni] = 0). If finite_infinite = false, the difference νi[ni-i+1]-νi[ni-i] for i = 1, 2, ..., ni is the number of infinite elementary divisors of degree i (with νi[0] = 0).

The full column rank pencil Ml-λNl, in a staircase form, contains the left Kronecker indices of M-λN and has the form

               | Al-λEl |
     Ml-λNl  = |--------|,
               |   Cl   |

where El is upper triangular and nonsingular. The nl-dimensional vectors νl and μl contain the row and, respectively, column dimensions of the blocks of the staircase form Ml-λNl such that the j-th block has dimensions νl[j] x μl[j] and has full column rank. The difference νl[nl-j+1]-μl[nl-j+1] for j = 1, 2, ..., nl is the number of elementary Kronecker blocks of size j x (j-1).

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of M, the absolute tolerance for the nonzero elements of N, and the relative tolerance for the nonzero elements of M and N. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed left orthogonal or unitary transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed right orthogonal or unitary transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

MatrixPencils.sklf_right — Function

sklf_right(A, E, B, C, D; fast = true, atol1 = 0, atol2 = 0, rtol, withQ = true, withZ = true) -> (F, G, Q, Z, νr, μr, nf, ν, μ)

Reduce the structured linear pencil M - λN

           | A-λE | B | 
  M - λN = |------|---|
           |  C   | D |

to an equivalent form F - λG = Q'*(M - λN)*Z using orthogonal or unitary transformation matrices Q and Z such that the transformed matrices F and G are in the following Kronecker-like form exhibiting the right and finite Kronecker structures:

            |  Mr-λNr    |     *      |    *     |
            |------------|------------|----------|
  F - λG =  |    O       |   Mf-λNf   |    *     |
            |------------|------------|----------|
            |    O       |     0      | Mli-λNli |

The full row rank pencil Mr-λNr, in a staircase form, contains the right Kronecker indices of M-λN and has the form

     Mr-λNr  = | Br | Ar-λEr |,

where Er is upper triangular and nonsingular. The nr-dimensional vectors νr and μr contain the row and, respectively, column dimensions of the blocks of the staircase form Mr-λNr such that i-th block has dimensions νr[i] x μr[i] and has full row rank. The difference μr[i]-νr[i] for i = 1, 2, ..., nr is the number of elementary Kronecker blocks of size (i-1) x i.

The nf x nf pencil Mf-λNf is regular and contains the finite elementary divisors of M-λN. Nf is upper triangular and nonsingular.

The full column rank pencil Mli-λNli is in a staircase form, and contains the left Kronecker indices and infinite elementary divisors of the pencil M-λN. The nb-dimensional vectors ν and μ contain the row and, respectively, column dimensions of the blocks of the staircase form Mli-λNli such that the i-th block has dimensions ν[i] x μ[i] and has full column rank. The difference ν[nb-j+1]-μ[nb-j+1] for j = 1, 2, ..., nb is the number of elementary Kronecker blocks of size j x (j-1). The difference μ[nb-j+1]-ν[nb-j] for j = 1, 2, ..., nb is the number of infinite elementary divisors of degree j (with ν[0] = 0).

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of M, the absolute tolerance for the nonzero elements of N, and the relative tolerance for the nonzero elements of M and N. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed left orthogonal or unitary transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed right orthogonal or unitary or unitary transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

Note: If the pencil M - λN has full row rank, then the regular pencil Mli-λNli is in a staircase form with square upper triangular diagonal blocks (i.e.,μ[j] = ν[j]), and the difference ν[nb-j+1]-ν[nb-j] for j = 1, 2, ..., nb is the number of infinite elementary divisors of degree j (with ν[0] = 0).

MatrixPencils.sklf_left — Function

sklf_left(A, E, B, C, D; fast = true, atol1 = 0, atol2 = 0, rtol, withQ = true, withZ = true) -> (F, G, Q, Z, ν, μ, nf, νl, μl)

Reduce the structured linear pencil M - λN

           | A-λE | B | 
  M - λN = |------|---|
           |  C   | D |

to an equivalent form F - λG = Q'*(M - λN)*Z using orthogonal or unitary transformation matrices Q and Z such that the transformed matrices F and G are in the following Kronecker-like form exhibiting the finite and left Kronecker structures:

            |  Mri-λNri  |     *      |    *    |
            |------------|------------|---------|
  F - λG =  |    O       |   Mf-λNf   |    *    |
            |------------|------------|---------|
            |    O       |     0      |  Ml-λNl |

The full row rank pencil Mri-λNri is in a staircase form, and contains the right Kronecker indices and infinite elementary divisors of the pencil M-λN.

The nf x nf pencil Mf-λNf is regular and contains the finite elementary divisors of M-λN. Nf is upper triangular and nonsingular.

The full column rank pencil Ml-λNl, in a staircase form, contains the left Kronecker indices of M-λN and has the form

               | Al-λEl |
     Ml-λNl  = |--------|,
               |   Cl   |

where El is upper triangular and nonsingular.

The nb-dimensional vectors ν and μ contain the row and, respectively, column dimensions of the blocks of the staircase form Mri-λNri such that i-th block has dimensions ν[i] x μ[i] and has full row rank. The difference μ[i]-ν[i] for i = 1, 2, ..., nb is the number of elementary Kronecker blocks of size (i-1) x i. The difference ν[i]-μ[i+1] for i = 1, 2, ..., nb is the number of infinite elementary divisors of degree i (with μ[nb+1] = 0). The nl-dimensional vectors νl and μl contain the row and, respectively, column dimensions of the blocks of the staircase form Ml-λNl such that the j-th block has dimensions νl[j] x μl[j] and has full column rank. The difference νl[nl-j+1]-μl[nl-j+1] for j = 1, 2, ..., nl is the number of elementary Kronecker blocks of size j x (j-1).

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of M, the absolute tolerance for the nonzero elements of N, and the relative tolerance for the nonzero elements of M and N. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed left orthogonal or unitary transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed right orthogonal or unitary transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

Note: If the pencil M - λN has full column rank, then the regular pencil Mri-λNri is in a staircase form with square upper triangular diagonal blocks (i.e.,μ[i] = ν[i]), and the difference ν[i+1]-ν[i] for i = 1, 2, ..., nb is the number of infinite elementary divisors of degree i (with ν[nb+1] = 0).

MatrixPencils.sreduceBF — Function

sreduceBF(A, E, B, C, D; fast = true, atol = 0, rtol, withQ = true, withZ = true) -> F, G, Q, Z, n, m, p

Reduce the partitioned linear pencil M - λN

           | A-λE | B | ndx
  M - λN = |------|---|
           |  C   | D | ny
              nx    nu

to an equivalent basic form F - λG = Q'*(M - λN)*Z using orthogonal transformation matrices Q and Z such that M - λN is transformed into the following standard form

           | B1 | A1-λE1 | n
  F - λG = |----|--------|   ,        
           | D1 |   C1   | p
             m      n

where E1 is an nxn non-singular matrix, and A1, B1, C1, D1 are nxn-, nxm-, pxn- and pxm-dimensional matrices, respectively. The order n of E1 is equal to the numerical rank of E determined using the absolute tolerance atol and relative tolerance rtol. The dimensions m and p are computed as m = nu + (nx-n) and p = ny + (ndx-n).

The performed orthogonal or unitary transformations are accumulated in Q, if withQ = true, and Z, if withZ = true.

If fast = true, E1 is determined upper triangular using a rank revealing QR-decomposition with column pivoting of E and n is evaluated as the number of nonzero diagonal elements of the R factor, whose magnitudes are greater than tol = max(atol,abs(R[1,1])*rtol). If fast = false, E1 is determined diagonal using a rank revealing SVD-decomposition of E and n is evaluated as the number of singular values greater than tol = max(atol,smax*rtol), where smax is the largest singular value. The rank decision based on the SVD-decomposition is generally more reliable, but the involved computational effort is higher.

MatrixPencils.sklf_right! — Function

sklf_right!(A, E, B, F, C, G, D, H; fast = true, atol1 = 0, atol2 = 0, atol3 = 0, rtol, withQ = true, withZ = true) -> (Q, Z, nc)

Reduce the partitioned full row rank linear pencil

  M - λN = | B-λF | A-λE | n
              m      n

with A-λE regular, to an equivalent form Mt - λNt = Q'*(M - λN)*Z = | Bt-λFt | At-λEt | using orthogonal or unitary transformation matrices Q and Z, such that M - λN is transformed into a Kronecker-like form | Bt-λFt | At-λEt | exhibiting its right, finite and infinite Kronecker structures (also known as the strong controllability form):

                         | Bc-λFc | Ac-λEc     *     | nc
  | Bt-λFt | At-λEt | =  |--------|------------------|
                         |   0    |   0     Auc-λEuc | n-nc
                             m        nc      n-nc

The matrices Bt, Ft, At, Et, determined from [Bt At] = Q'*[B A]*Z, [Ft Et] = Q'*[F E]*Z, are returned in B, F, A and E, respectively. Furthermore, the matrices Ct, Dt, Gt, Ht, determined from the compatibly partitioned matrices [Dt Ct] = [D C]*Z and [Ht Gt] = [D C]*Z, are returned in C, D, G and H, respectively (unless C, D',GandH` are set to 'missing').

The subpencil | Bc-λFc | Ac-λEc | has full row rank nc and the (n-nc) x (n-nc) subpencil Auc-λEuc contains the finite and infinite eigenvalues of M - λN (also called the uncontrollable eigenvalues of M - λN).

The (m+n) x (m+n) orthogonal matrix Z has the partitioned form

       | Z11 |  0  |  *  | m 
   Z = |-----|-----|-----|
       |  *  |  *  |  *  | n
          m    nc    n-nc

with the leading m x m submatrix Z11 invertible and upper triangular.

Note: If Ct-λGt is partitioned as [ Cc-λGc | Cuc-λGuc ], then the following relation is fulfilled

  (Cc-λGc)*inv(λEc-Ac)(Bc-λFc) + Dt-λHt = ((C-λG)*inv(λE-A)(B-λF) + D-λH)*Z11

and the structured pencil linearization (Ac-λEc,Bc-λFc,Cc-λGc,Dt-λHt) is strongly controllable.

The keyword arguments atol1, atol2, , atol3, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A and B, the absolute tolerance for the nonzero elements of E and F, and the relative tolerance for the nonzero elements of A, B, E and F. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed orthogonal or unitary left transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed orthogonal or unitary right transformations are accumulated in the matrix Z if withZ = true. If withZ = false, Z contains the upper triangular matrix Z11.

sklf_right!(A, E, B, C; fast = true, atol1 = 0, atol2 = 0, atol3 = 0, rtol, withQ = true, withZ = true) -> (Q, Z, νr, nc, nfuc, niuc)

Reduce the partitioned full row rank linear pencil

  M - λN = | B | A-λE | n
             m    n

with A-λE regular, to an equivalent form F - λG = Q'*(M - λN)*diag(I,Z) using orthogonal or unitary transformation matrices Q and Z such that M - λN is transformed into a Kronecker-like form | Bt | At-λEt | exhibiting its right, finite and infinite Kronecker structures (also known as the generalized controllability staircase form):

                     |  Bc | Ac-λEc     *          *      | nc
  | Bt | At-λEt | =  |-----|------------------------------|
                     |  0  |  0     Afuc-λEfuc     *      | nfuc
                     |  0  |  0         0      Aiuc-λEiuc | niuc
                        m     nc       nfuc        niuc

Bt = Q'*B, At = Q'*A*Zand Et = Q'*E*Z are returned in B, A and E, respectively, and C*Z is returned in C (unless `C = missing').

The subpencil | Bc | Ac-λEc | has full row rank nc, is in a staircase form, and contains the right Kronecker indices of M - λN. The nr-dimensional vector νr contains the row and column dimensions of the blocks of the staircase form | Bc | Ac-λI | such that i-th block has dimensions νr[i] x νr[i-1] (with νr[0] = m) and has full row rank. The difference νr[i-1]-νr[i] for i = 1, 2, ..., nr is the number of elementary Kronecker blocks of size (i-1) x i.

The nfuc x nfuc subpencil Afuc-λEfuc contains the finite eigenvalues of M - λN (also called the uncontrollable finite eigenvalues of A - λE).

The niuc x niuc subpencil Aiuc-λEiuc contains the infinite eigenvalues of M - λN (also called the uncontrollable infinite eigenvalues of A - λE).

The keyword arguments atol1, atol2, , atol3, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, the absolute tolerance for the nonzero elements of B, and the relative tolerance for the nonzero elements of A, E and B. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed orthogonal or unitary left transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed orthogonal or unitary right transformations are accumulated in the matrix Z if withZ = true or C is provided. Otherwise, Z is set to nothing.

sklf_right!(A, B, C; fast = true, atol1 = 0, atol2 = 0, rtol, withQ = true) -> (Q, νr, nc, nuc)

Reduce the partitioned full row rank linear pencil

  M - λN = | B | A-λI | n
             m    n

to an equivalent form F - λG = Q'*(M - λN)*diag(I,Q) using orthogonal or unitary transformation matrix Q such that M - λN is transformed into a Kronecker-like form | Bt | At-λI | exhibiting its right and finite Kronecker structures (also known as the controllability staircase form):

                    |  Bc | Ac-λI     *    | nc
  | Bt | At-λI | =  |-----|----------------|
                    |  0  |  0     Auc-λI  | nuc
                       m     nc      nuc

Bt = Q'*B and At = Q'*A*Q are returned in B and A, respectively, and C*Q is returned in C (unless `C = missing').

The subpencil | Bc | Ac-λI | has full row rank nc, is in a staircase form, and contains the right Kronecker indices of M - λN. The nr-dimensional vector νr contains the row and column dimensions of the blocks of the staircase form | Bc | Ac-λI | such that i-th block has dimensions νr[i] x νr[i-1] (with νr[0] = m) and has full row rank. The difference νr[i-1]-νr[i] for i = 1, 2, ..., nr is the number of elementary Kronecker blocks of size (i-1) x i.

The nuc x nuc matrix Auc contains the (finite) eigenvalues of M - λN (also called the uncontrollable eigenvalues of A).

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of B, and the relative tolerance for the nonzero elements of A and B. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed orthogonal or unitary transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing.

MatrixPencils.sklf_left! — Function

sklf_left!(A, E, C, B, F, G, D, H; fast = true, atol1 = 0, atol2 = 0, atol3 = 0, rtol, withQ = true, withZ = true) -> (Q, Z, no)

Reduce the partitioned full row rank linear pencil

M - λN = | A-λE | n
         | C-λG | p
            n

with A-λE regular, to an equivalent form F - λG = Q'*(M - λN)*Z using orthogonal or unitary transformation matrices Q and Z such that M - λN is transformed into a staircase form exhibiting the left, finite and infinite Kronecker structures (also known as the strong observability form):

                | Auo-λEuo  |   *    | n-no
  | At-λEt |    |   0       | Ao-λEo | no
  |--------| =  |-----------|--------|
  | Ct-λGt |    |   0       | Co-λGo | p
                   n-no         no

The matrices Ct, Gt, At, Et, determined from [At; Ct] = Q'*[A; C]*Z, [Et; Gt] = Q'*[E; G]*Z, are returned in C, G, A and E, respectively. Furthermore, the matrices Bt, Dt, Ft, Ht determined from the compatibly partitioned [Bt; Dt] = Q'*[B; D] and [Ft; Ht] = Q'*[F; H], are returned in B, D, F, and H, respectively (unless B, D',FandH` are set to 'missing').

The subpencil

   | Ao-λEo |   
   |--------|
   | Co-λGo |

has full column rank no and the (n-no) x (n-no) subpencil Auo-λEuo contains the finite and infinite eigenvalues of M - λN (also called the unobservable eigenvalues of M - λN).

The (n+p) x (n+p) orthogonal matrix Q has the partitioned form

       |  *  |  *  |  *   | n 
   Q = |-----|-----|------|
       |  *  |  0  | Q22  | p
         n-no  no     p

with the p x p trailing submatrix Q22 invertible and upper triangular.

Note: If Bt-λFt is partitioned as [ Buo-λFuo | Bo-λFo ], then the following relation is fulfilled

  (Co-λGo)*inv(λEo-Ao)(Bo-λFo) + Dt-λHt = Q22'*((C-λG)*inv(λE-A)(B-λF) + D-λH)

and the structured pencil linearization (Ao-λEo,Bo-λFo,Co-λGo,Dt-λHt) is strongly observable.

The keyword arguments atol1, atol2, , atol3, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A and C, the absolute tolerance for the nonzero elements of E and G, and the relative tolerance for the nonzero elements of A, C, E and G. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false. The performed orthogonal or unitary left transformations are accumulated in the matrix Q if withQ = true. If withQ = false, Q contains the upper triangular matrix Q22. The performed orthogonal or unitary right transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

sklf_left!(A, E, C, B; fast = true, atol1 = 0, atol2 = 0, atol3 = 0, rtol, withQ = true, withZ = true) -> (Q, Z, μl, no, nfuo, niuo)

Reduce the partitioned full column rank linear pencil

  M - λN = | A-λE | n
           |  C   | p
              n

to an equivalent form F - λG = diag(Q',I)*(M - λN)*Z using orthogonal or unitary transformation matrices Q and Z such that M - λN is transformed into a Kronecker-like form exhibiting its infinite, finite and left Kronecker structures (also known as the generalized observability staircase form):

                | Aiuo-λEiuo     *       |   *    | niuo
                |    0       Afuo-λEfuo  |   *    | nfuo
  | At-λEt |    |    0           0       | Ao-λEo | no
  |--------| =  |------------------------|--------|
  |  Ct    |    |    0           0       |   Co   | p
                    niuo        nfuo         no

Ct = C*Z, At = Q'*A*Z and Et = Q'*E*Z are returned in C, A and E, respectively, and Q'*B is returned in B (unless `B = missing').

The subpencil | Ao-λEo | has full column rank no, is in a staircase form, and contains the left Kronecker indices of M - λN. | Co | The nl-dimensional vector μl contains the row and column dimensions of the blocks of the staircase form such that i-th block has dimensions μl[nl-i] x μl[nl-i+1] (with μl[0] = p) and has full column rank. The difference μl[nl-i]-μl[nl-i+1] for i = 1, 2, ..., nl is the number of elementary Kronecker blocks of size i x (i-1).

The niuo x niuo subpencil Aiuo-λEiuo contains the infinite eigenvalues of M - λN (also called the unobservable infinite eigenvalues of A-λE).

The nfuo x nfuo subpencil Afuo-λEfuo contains the finite eigenvalues of M - λN (also called the unobservable finite eigenvalues of A-λE).

The keyword arguments atol1, atol2, atol3 and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of E, the absolute tolerance for the nonzero elements of C, and the relative tolerance for the nonzero elements of A, E and C. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed orthogonal or unitary left transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed orthogonal or unitary right transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

sklf_left!(A, C, B; fast = true, atol1 = 0, atol2 = 0, rtol, withQ = true) -> (Q, μl, no, nuo)

Reduce the partitioned full column rank linear pencil

  M - λN = | A-λI | n
           |  C   | p
              n

to an equivalent form F - λL = diag(Q',I)*(M - λN)*Q using orthogonal or unitary transformation matrix Q such that M - λN is transformed into a Kronecker-like form exhibiting its finite and left Kronecker structures (also known as the observability staircase form):

               | Auo-λI  |  *    | nuo
  | At-λI |    |    0    | Ao-λI | no
  |-------| =  |---------|-------|
  |  Ct   |    |    0    |   Co  | p
                   nuo       no

Ct = C*Q and At = Q'*A*Q are returned in C and A, respectively, and Q'*B is returned in B (unless `B = missing').

The subpencil | Ao-λI; Co | has full column rank no, is in a staircase form, and contains the left Kronecker indices of M - λN.

The nl-dimensional vector μl contains the row and column dimensions of the blocks of the staircase form such that i-th block has dimensions μl[nl-i] x μl[nl-i+1] (with μl[0] = p) and has full column rank. The difference μl[nl-i]-μl[nl-i+1] for i = 1, 2, ..., nl is the number of elementary Kronecker blocks of size i x (i-1).

The nuo x nuo matrix Auo contains the (finite) eigenvalues of M - λN (also called the unobservable eigenvalues of A).

The keyword arguments atol1, atol2, and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of C, and the relative tolerance for the nonzero elements of A and C. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed orthogonal or unitary transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing.

MatrixPencils.sklf_rightfin! — Function

sklf_rightfin!(A, E, B, C; fast = true, atol1 = 0, atol2 = 0,  
               rtol, withQ = true, withZ = true) -> (Q, Z, νr, nc, nuc)

Reduce the partitioned full row rank linear pencil

  M - λN = | B | A-λE | n
             m    n

with A-λE regular, to an equivalent form F - λG = Q'*(M - λN)*diag(I,Z) using orthogonal or unitary transformation matrices Q and Z such that M - λN is transformed into a staircase form | Bt | At-λEt | exhibiting the separation of its finite eigenvalues:

                     |  Bc | Ac-λEc     *    | nc
  | Bt | At-λEt | =  |-----|-----------------|
                     |  0  |  0     Auc-λEuc | nuc
                        m     nc      nuc

Bt = Q'*B, At = Q'*A*Zand Et = Q'*E*Z are returned in B, A and E, respectively, and C*Z is returned in C (unless C = missing'). The resultingEtis upper triangular. IfEis already upper triangular, then the preliminary reduction ofE` to upper triangular form is not performed.

The subpencil | Bc | Ac-λEc | has full row rank nc for all finite values of λ, is in a staircase form, and, if E is invertible, contains the right Kronecker indices of M - λN. The nr-dimensional vector νr contains the row and column dimensions of the blocks of the staircase form | Bc | Ac-λI | such that i-th block has dimensions νr[i] x νr[i-1] (with νr[0] = m) and has full row rank. If E is invertible, the difference νr[i-1]-νr[i] for i = 1, 2, ..., nr is the number of elementary Kronecker blocks of size (i-1) x i.

The nuc x nuc subpencil Auc-λEuc contains the finite eigenvalues of M - λN (also called the uncontrollable finite eigenvalues of A - λE). If E is singular, Auc-λEuc may also contain a part of the infinite eigenvalues of M - λN (also called the uncontrollable infinite eigenvalues of A - λE).

The keyword arguments atol1, atol2 and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of B, and the relative tolerance for the nonzero elements of A and B. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed orthogonal or unitary left transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed orthogonal or unitary right transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

Note: This function, called with reversed input parameters E and A (i.e., instead A and E), performs the separation all infinite and nonzero finite eigenvalues of the pencil M - λN.

MatrixPencils.sklf_leftfin! — Function

sklf_leftfin!(A, E, C, B; fast = true, atol1 = 0, atol2 = 0,  
              rtol, withQ = true, withZ = true) -> (Q, Z, μl, no, nuo)

Reduce the partitioned full column rank linear pencil

  M - λN = | A-λE | n
           |  C   | p
              n

with A-λE regular, to an equivalent form F - λG = Q'*(M - λN)*diag(I,Z) using orthogonal or unitary transformation matrices Q and Z such that M - λN is transformed into a staircase form exhibiting the separation of its finite eigenvalues:

                | Auo-λEuo  |   *    | nuo
  | At-λEt |    |   0       | Ao-λEo | no
  |--------| =  |-----------|--------|
  |  Ct    |    |   0       |   Co   | p
                    nuo         no

Ct = C*Z, At = Q'*A*Z and Et = Q'*E*Z are returned in C, A and E, respectively, and Q'*B is returned in B (unless B = missing'). The resultingEtis upper triangular. IfEis already upper triangular, then the preliminary reduction ofE` to upper triangular form is not performed.

The subpencil | Ao-λEo | has full column rank no for all finite values of λ, is in a staircase form, | Co | and, if E is nonsingular, contains the left Kronecker indices of M - λN. The nl-dimensional vector μl contains the row and column dimensions of the blocks of the staircase form such that i-th block has dimensions μl[nl-i] x μl[nl-i+1] (with μl[0] = p) and has full column rank. If E is nonsingular, the difference μl[nl-i]-μl[nl-i+1] for i = 1, 2, ..., nl is the number of elementary Kronecker blocks of size i x (i-1).

The keyword arguments atol1, atol2 and rtol, specify, respectively, the absolute tolerance for the nonzero elements of A, the absolute tolerance for the nonzero elements of C, and the relative tolerance for the nonzero elements of A and C. The reduction is performed using rank decisions based on rank revealing QR-decompositions with column pivoting if fast = true or the more reliable SVD-decompositions if fast = false.

The performed orthogonal or unitary left transformations are accumulated in the matrix Q if withQ = true. Otherwise, Q is set to nothing. The performed orthogonal or unitary right transformations are accumulated in the matrix Z if withZ = true. Otherwise, Z is set to nothing.

Note: This function, called with reversed input parameters E and A (i.e., instead A and E), performs the separation all infinite and nonzero finite eigenvalues of the pencil M - λN.